requirements.txt
python-telegram-bot==21.9
requests==2.31.0
python-dotenv==1.0.0
aiosqlite==0.19.0
Pillow>=11.0.0
httpx>=0.28.0